+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add the possibility to select a folder in the function `count_seq()`
 * Add functions `track_wkflow_samples()` and `select_one_sample()`
+* Add option `sam_data_first` in function `write_pq()`
 
 # MiscMetabar 0.31 (in development)